[feat] home fetch 로직을 Repository를 통해 호출하도록 구현
