+++ Routinus/Routinus/Data/RoutinusRepository.swift
@@ -7,31 +7,10 @@
 
 import Foundation
 
-import RoutinusDatabase
-
-protocol HomeRepository {
-    func isEmptyUserID() -> Bool
-    func save(id: String, name: String)
-}
-
 class RoutinusRepository {
     static let userIDKey = "id"
-    
+
     static func userID() -> String? {
         return UserDefaults.standard.string(forKey: RoutinusRepository.userIDKey)
-    }
-}
-
-extension RoutinusRepository: HomeRepository {
-    func isEmptyUserID() -> Bool {
-        return UserDefaults.standard.string(forKey: RoutinusRepository.userIDKey) == nil
-    }
-
-    func save(id: String, name: String) {
-        UserDefaults.standard.set(id, forKey: RoutinusRepository.userIDKey)
-        
-        Task {
-            try await RoutinusDatabase.createUser(id: id, name: name)
-        }
     }
 }
